refactor: 컴포넌트 변수명 변
