+++ src/layouts/BaseLayout/Footer/styles.ts
@@ -1,0 +1,5 @@
+import { css } from '@emotion/react';
+
+export const footerStyles = css({
+  padding: '2rem 0',
+});
