+++ src/layouts/BaseLayout/Header/styles.ts
@@ -1,0 +1,7 @@
+import { css } from '@emotion/react';
+
+export const headerStyle = css({
+  position: 'fixed',
+  top: '0',
+  zIndex: '100',
+});
