+++ src/pages/HomePage/components/RankSection/RankFilterUser/index.tsx
@@ -1,4 +1,4 @@
-import { UserFilterButtons } from '@/data/filterData';
+import { UserContents } from '@/data/filterButton';
 
 import { Container } from '@/components/ui/Layout/Container';
 
@@ -7,7 +7,7 @@ import { ActiveUserButton } from './ActiveUserButton';
 export const RankFilterUser = () => {
   return (
     <Container justifyContent="space-around">
-      {UserFilterButtons.map((filter) => (
+      {UserContents.map((filter) => (
         <ActiveUserButton key={filter.index} user={filter} />
       ))}
     </Container>
